pyproject.toml
[tool.poetry]
name = "agentflow"
version = "0.1.0"
description = "A scalable AI Agent workflow framework with YAML-first design and visual editor"
authors = ["ddLowell <ddlowell@example.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "agentflow"}]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
httpx = "^0.25.0"
typer = {extras = ["all"], version = "^0.9.0"}
pydantic-settings = "^2.1.0"
aiofiles = "^23.2.0"
python-multipart = "^0.0.6"

# MCP Protocol
mcp = "^0.9.0"

# Database and Vector DB
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"
ruff = "^0.1.6"

[tool.poetry.scripts]
agentflow = "agentflow.cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]